vers 1.12

+++ ScadenzaDiLegge/Delegate/DeleteDelegate.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 
 namespace ScadenzaDiLegge.Delegate
 {
@@ -11,8 +12,24 @@ namespace ScadenzaDiLegge.Delegate
         delegate void deleteDelegate();
        public static void DeleteDel()
         {
-            DeleteWindow deleteWindow = new DeleteWindow();
-            deleteWindow.Show();
+            DeleteWindow deleteWindow = Application.Current.Windows
+     .OfType<DeleteWindow>()
+     .FirstOrDefault();
+
+            if (deleteWindow == null)
+            {
+                deleteWindow = new DeleteWindow();
+                deleteWindow.Show();
+            }
+            else
+            {
+                if (deleteWindow.WindowState == WindowState.Minimized)
+                    deleteWindow.WindowState = WindowState.Normal;
+
+                deleteWindow.Focus();
+                deleteWindow.Activate();
+            }
+
         }
     }
 }
